Move ConnectProtocolConnectionDecorator
